refactor(cms): move `topNTopicsOfLegislators` typedefs and resolver to extend-graphql-schemas/top-n-topics-of-legislators.ts
